+++ src/meltano/cli/install.py
@@ -2,7 +2,7 @@ import click
 from . import cli
 from meltano.core.plugin_install_service import PluginInstallService
 from meltano.core.plugin_discovery_service import PluginNotFoundError
-from meltano.core.project import Project
+from meltano.core.project import Project, ProjectNotFound
 from meltano.core.project_add_service import ProjectAddService
 
 
@@ -23,7 +23,11 @@ def install_status_update(data):
 
 @cli.command()
 def install():
-    project = Project.find()
+    try:
+        project = Project.find()
+    except ProjectNotFound as e:
+        raise click.ClickException(e)
+
     install_service = PluginInstallService(project)
     install_status = install_service.install_all_plugins(install_status_update)
     num_installed = len(install_status["installed"])