+++ CHANGELOG.md
@@ -17,6 +17,19 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## 0.3.2 - (2018-12-21)
+---
+
+## 0.3.1 - (2018-12-21)
+---
+
+### Changes
+* add default models for 'tap-carbon-intensity'.
+* Meltano Analyze is now part of the package.
+* removes database dependency from Meltano Analyze and uses .ma files
+* update the error message when using Meltano from outside a project - [238](https://gitlab.com/meltano/meltano/merge_requests/238)
+
+
 ## 0.3.0 - (2018-12-18)
 ---